+++ apps/common/middleware/cross_domain_middleware.py
@@ -6,22 +6,10 @@
     @date：2024/5/8 13:36
     @desc:
 """
-from django.db.models import QuerySet
 from django.http import HttpResponse
 from django.utils.deprecation import MiddlewareMixin
 
-from application.models.api_key_model import ApplicationApiKey
-from common.constants.cache_code_constants import CacheCodeConstants
-from common.util.cache_util import get_cache
-
-
-@get_cache(cache_key=lambda secret_key, use_get_data: secret_key,
-           use_get_data=lambda secret_key, use_get_data: use_get_data,
-           version=CacheCodeConstants.APPLICATION_API_KEY_CACHE.value)
-def get_application_api_key(secret_key, use_get_data):
-    application_api_key = QuerySet(ApplicationApiKey).filter(secret_key=secret_key).first()
-    return {'allow_cross_domain': application_api_key.allow_cross_domain,
-            'cross_domain_list': application_api_key.cross_domain_list}
+from common.cache_data.application_api_key_cache import get_application_api_key
 
 
 class CrossDomainMiddleware(MiddlewareMixin):
